Cloud 3.1.0

+++ components/device-mgt/org.wso2.carbon.device.mgt.api/src/main/java/org/wso2/carbon/device/mgt/jaxrs/beans/BasePaginatedResult.java
@@ -30,13 +30,13 @@ public class BasePaginatedResult {
             value = "Number of total resources.",
             example = "1")
     @JsonProperty("count")
-    private long count;
+    private int count;
 
-    public long getCount() {
+    public int getCount() {
         return count;
     }
 
-    public void setCount(long count) {
+    public void setCount(int count) {
         this.count = count;
     }
 }
